KS-1-4 4th task solution corrected

+++ src/main/kotlin/lesson_1/lesson1_task4.kt
@@ -4,7 +4,7 @@ fun main(){
     val distance: Long = 40_868_600_000L
     val age: Byte = 27
     val partOfDay: Float = 0.075F
-    val sec: Int = 6480
+    val sec: Short = 6480
     val partOfYear: Double = 2.0547945205479453E-4
     val apogee: Int = 327000
